@@ -1,13 +1,23 @@
 module github.com/shimmeringbee/zstack
 
-go 1.14
+go 1.22.0
+
+toolchain go1.22.2
 
 require (
 	github.com/shimmeringbee/bytecodec v0.0.0-20210228205504-1e9e0677347b
 	github.com/shimmeringbee/logwrap v0.1.3
+	github.com/shimmeringbee/persistence v0.0.0-20240521204303-bf4ab8a6b71b
 	github.com/shimmeringbee/retry v0.0.0-20221006193055-2ce01bf139c2
 	github.com/shimmeringbee/unpi v0.0.0-20210525151328-7ede275a1033
-	github.com/shimmeringbee/zigbee v0.0.0-20221016122511-6c2328db0d94
+	github.com/shimmeringbee/zigbee v0.0.0-20240614090423-d67fd427d102
 	github.com/stretchr/testify v1.9.0
 	golang.org/x/sync v0.7.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/objx v0.5.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
