@@ -3,6 +3,7 @@ package zstack
 import (
 	"context"
 	"github.com/shimmeringbee/bytecodec"
+	"github.com/shimmeringbee/persistence/impl/memory"
 	. "github.com/shimmeringbee/unpi"
 	unpiTest "github.com/shimmeringbee/unpi/testing"
 	"github.com/shimmeringbee/zigbee"
@@ -18,7 +19,7 @@ func Test_ReceiveMessage(t *testing.T) {
 
 		unpiMock := unpiTest.NewMockAdapter()
 		defer unpiMock.AssertCalls(t)
-		zstack := New(unpiMock)
+		zstack := New(unpiMock, memory.New())
 		defer unpiMock.Stop()
 
 		zstack.nodeTable.addOrUpdate(zigbee.IEEEAddress(0x1122334455667788), zigbee.NetworkAddress(0x1000))
